Fixed our problem. If you want to test comment the sending of acks in MClientListener.
Some problems where that the message was never removed after retries.
Also there was a check on username for removing messages this is
replaced by a check on processIDs (pids are also part of msgText now).
Hope this last part did not break anything.

Next up is to check thoroughly if the two ack systems are fully working
before going on to the next one.