COOL Dynamic Load balancing, brings tears to my eyes (of joy)
Master monitors the load of clients over hosts. If it detects an
unbalanced load it orders hosts which order clients to reroute to a host
which has the least nr of clients. 

Multiple clients can be rerouted at the same time (for example one host
with 4 clients and other host with 0 clients will become 2 and 2).

Works on clients joining/leaving and hosts joining

Users do not notice anything but a change in the port number