Client reconnects when host dies.
No heartbeats needed (code is there but commented) as a detection that a
connection is closed between client and host can be detected. 

The connect and send message buttons are disabled when appropriate. This
to prevent bugs from occuring. (Could not fix a bug when trying to
reconnect, then send message, then connected, then no messages are
delivered anymore).